norminized lexer, log, loop, main
