+++ src/loop.c
@@ -6,7 +6,7 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/03 15:59:50 by aprevrha          #+#    #+#             */
-/*   Updated: 2024/05/03 15:32:58 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/07 12:40:20 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,13 +60,11 @@ int	process_line(char *line, t_evars *evars)
 int	shell_loop(t_list *env_list, int tty)
 {
 	char	*line;
-	t_evars evars;
+	t_evars	evars;
 
-	
 	evars.status = 0;
 	evars.tty = tty;
 	evars.envp = env_list;
-	
 	while (1)
 	{
 		idle_signals();
