+++ src/main.c
@@ -6,17 +6,16 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/01 15:06:11 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 19:22:46 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/07 12:34:55 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-
 void	clear_gnl_buf(void)
 {
 	char	*str;
-	
+
 	str = NULL;
 	log_msg(DEBUG, "running gnl until it is NULL");
 	str = get_next_line(STDIN_FILENO);
@@ -34,7 +33,6 @@ int	main(int argc, char const *argv[], char *envp[])
 	int		exit_status;
 	int		tty;
 
-
 	(void)argc;
 	(void)argv;
 	tty = isatty(STDIN_FILENO);
@@ -43,7 +41,6 @@ int	main(int argc, char const *argv[], char *envp[])
 	envp_lst = envp_to_list(envp);
 	if (!envp_lst)
 		return (FAILURE);
-	
 	exit_status = shell_loop(envp_lst, tty);
 	ft_lstclear(&envp_lst, free_env);
 	if (!tty)
