+++ src/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 16:41:35 by hrother           #+#    #+#             */
-/*   Updated: 2024/05/06 20:51:10 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/07 12:53:51 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,13 @@
 int	ft_strcmp(const char *s1, const char *s2)
 {
 	return (ft_strncmp(s1, s2, ft_strlen(s2) + 1));
+}
+
+void	skip_until(const char *str, unsigned int *i, const char *charset,
+		bool val)
+{
+	while ((ft_strchr(charset, str[*i]) != 0) != val && str[*i] != '\0')
+		*i += 1;
 }
 
 t_cmd	*new_cmd(char *bin, char **args)
